Redirect to home page when edited  work unit submitted after navigated to by URL. Finish [#107781324]

+++ app/controllers/work_units_controller.rb
@@ -109,8 +109,12 @@ class WorkUnitsController < WorkUnitBaseController
   end
 
   def redirect_back
-    redirect_to(session[:return_to])
-    session.delete(:return_to)
+    if session[:return_to]
+      redirect_to(session[:return_to])
+      session.delete(:return_to)
+    else
+      redirect_to root_path
+    end
   end
 
   def store_location
